feat: implement BaseResponse in RoomTypeFacilityDetailOption controller and update service

+++ src/main/java/com/tobeto/java4a/hotelnow/controllers/RoomTypeFacilityDetailOptionsController.java
@@ -1,12 +1,20 @@
 package com.tobeto.java4a.hotelnow.controllers;
 
+import com.tobeto.java4a.hotelnow.core.utils.messages.Messages;
+import com.tobeto.java4a.hotelnow.services.abstracts.RoomTypeFacilityCategoryService;
 import com.tobeto.java4a.hotelnow.services.abstracts.RoomTypeFacilityDetailOptionService;
 import com.tobeto.java4a.hotelnow.services.dtos.requests.roomtypefacilitydetailoptions.AddRoomTypeFacilityDetailOptionRequest;
 import com.tobeto.java4a.hotelnow.services.dtos.requests.roomtypefacilitydetailoptions.UpdateRoomTypeFacilityDetailOptionRequest;
+import com.tobeto.java4a.hotelnow.services.dtos.responses.BaseResponse;
+import com.tobeto.java4a.hotelnow.services.dtos.responses.roomtypefacilitycategories.ListRoomTypeFacilityCategoryResponse;
 import com.tobeto.java4a.hotelnow.services.dtos.responses.roomtypefacilitydetailoptions.AddRoomTypeFacilityDetailOptionResponse;
 import com.tobeto.java4a.hotelnow.services.dtos.responses.roomtypefacilitydetailoptions.ListRoomTypeFacilityDetailOptionResponse;
 import com.tobeto.java4a.hotelnow.services.dtos.responses.roomtypefacilitydetailoptions.UpdateRoomTypeFacilityDetailOptionResponse;
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
+import org.springframework.data.repository.query.Param;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,28 +22,47 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/room-type-facility-detail-options")
 @AllArgsConstructor
-public class RoomTypeFacilityDetailOptionsController {
+public class RoomTypeFacilityDetailOptionsController extends BaseController {
 
-    private RoomTypeFacilityDetailOptionService roomTypeFacilityDetailOptionService;
+    private final RoomTypeFacilityDetailOptionService roomTypeFacilityDetailOptionService;
+    private final RoomTypeFacilityCategoryService roomTypeFacilityCategoryService;
 
     @GetMapping("/by-category-id/{categoryId}")
-    public List<ListRoomTypeFacilityDetailOptionResponse> getByCategoryId(@PathVariable int categoryId) {
-        return roomTypeFacilityDetailOptionService.getByCategoryId(categoryId);
+    public ResponseEntity<BaseResponse<List<ListRoomTypeFacilityDetailOptionResponse>>> getByCategoryId(@PathVariable int categoryId) {
+        ListRoomTypeFacilityCategoryResponse roomTypeFacilityCategory = roomTypeFacilityCategoryService.getResponseById(categoryId);
+        List<ListRoomTypeFacilityDetailOptionResponse> roomTypeFacilityDetailOptions = roomTypeFacilityDetailOptionService.getByCategoryId(categoryId);
+        if (roomTypeFacilityCategory != null && roomTypeFacilityDetailOptions != null && !roomTypeFacilityDetailOptions.isEmpty()) {
+            return sendResponse(HttpStatus.OK, Messages.Success.CUSTOM_LISTED_SUCCESSFULLY, roomTypeFacilityDetailOptions);
+        }
+        return sendResponse(HttpStatus.NOT_FOUND, Messages.Error.CUSTOM_ROOM_TYPE_FACILITY_DETAIL_OPTION_NOT_FOUND, null);
     }
 
     @PostMapping
-    public AddRoomTypeFacilityDetailOptionResponse add(@RequestBody AddRoomTypeFacilityDetailOptionRequest request) {
-        return roomTypeFacilityDetailOptionService.add(request);
+    public ResponseEntity<BaseResponse<AddRoomTypeFacilityDetailOptionResponse>> add(@RequestBody @Valid AddRoomTypeFacilityDetailOptionRequest request) {
+        ListRoomTypeFacilityCategoryResponse roomTypeFacilityCategory = roomTypeFacilityCategoryService.getResponseById(request.getCategoryId());
+        if (roomTypeFacilityCategory != null) {
+            return sendResponse(HttpStatus.OK, Messages.Success.CUSTOM_CREATED_SUCCESSFULLY, roomTypeFacilityDetailOptionService.add(request));
+        }
+        return sendResponse(HttpStatus.NOT_FOUND, Messages.Error.CUSTOM_ROOM_TYPE_FACILITY_CATEGORY_NOT_FOUND, null);
     }
 
     @PutMapping
-    public UpdateRoomTypeFacilityDetailOptionResponse update(@RequestBody UpdateRoomTypeFacilityDetailOptionRequest request) {
-        return roomTypeFacilityDetailOptionService.update(request);
+    public ResponseEntity<BaseResponse<UpdateRoomTypeFacilityDetailOptionResponse>> update(@RequestBody @Valid UpdateRoomTypeFacilityDetailOptionRequest request) {
+        ListRoomTypeFacilityDetailOptionResponse roomTypeFacilityDetailOption = roomTypeFacilityDetailOptionService.getResponseById(request.getId());
+        if (roomTypeFacilityDetailOption != null) {
+            return sendResponse(HttpStatus.OK, Messages.Success.CUSTOM_UPDATED_SUCCESSFULLY, roomTypeFacilityDetailOptionService.update(request));
+        }
+        return sendResponse(HttpStatus.NOT_FOUND, Messages.Error.CUSTOM_ROOM_TYPE_FACILITY_DETAIL_OPTION_NOT_FOUND, null);
     }
 
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
-        roomTypeFacilityDetailOptionService.delete(id);
+    @DeleteMapping
+    public ResponseEntity<BaseResponse<String>> delete(@Param("id") int id) {
+        ListRoomTypeFacilityDetailOptionResponse roomTypeFacilityDetailOptionToBeFound = roomTypeFacilityDetailOptionService.getResponseById(id);
+        if (roomTypeFacilityDetailOptionToBeFound != null) {
+            roomTypeFacilityDetailOptionService.delete(id);
+            return sendResponse(HttpStatus.OK, Messages.Success.CUSTOM_DELETED_SUCCESSFULLY, null);
+        }
+        return sendResponse(HttpStatus.NOT_FOUND, Messages.Error.CUSTOM_ROOM_TYPE_FACILITY_DETAIL_OPTION_NOT_FOUND, null);
     }
 
 }
